test-requirements.txt
Django==1.8.3
factory-boy==2.5.2
selenium==2.46.0
wheel==0.24.0
